moved text-color-style-fn.js to common
