+++ examples/main/shared/preview/PreviewPlugins.jsx
@@ -27,7 +27,7 @@ import {
 import {
   viewerCustomForegroundStyleFn,
   styleSelectionPredicate,
-} from '../../src/text-color-style-fn';
+} from 'wix-rich-content-common/dist/lib/text-color-style-fn';
 
 import 'wix-rich-content-editor-common/dist/styles.min.css';
 import 'wix-rich-content-common/dist/styles.min.css';
