+++ packages/common/web/src/lib/text-color-style-fn.js
